pyproject.toml
[tool.poetry]
name = "dotflow-api"
version = "0.1.0"
description = "Dotflow API"
package-mode = false
authors = ["FernandoCelmer <email@fernandocelmer.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = "^2.11.2"
pyjwt = "^2.10.1"
six = "^1.17.0"
opentelemetry-instrumentation-fastapi = "^0.57b0"
sqlalchemy = "^2.0.42"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
